feat: 회의 생성 페이지 및 관련 훅 리팩토링, 템플릿 및 제안 목록 추가
