+++ src/components/meeting/create/entities/MeetingTextArea.tsx
@@ -1,3 +1,5 @@
+import { TEMPLATES } from "@/mocks/meeting/create/templates";
+
 export default function MeetingTextArea({ value, onChange }) {
   const template = ({ text, onClick }) => {
     return <button onClick={onClick}>{text}</button>;
@@ -23,22 +25,3 @@ export default function MeetingTextArea({ value, onChange }) {
     </div>
   );
 }
-
-export const TEMPLATES = [
-  {
-    text: "안건 & 논의사항.",
-    onClick: () => console.log("안건 선택"),
-  },
-  {
-    text: " 녹음 파일로 회의록 작성하기",
-    onClick: () => console.log("녹음 파일로 회의록 작성"),
-  },
-  {
-    text: "신입생 OT (예시)",
-    onClick: () => console.log("신입생 OT 예시"),
-  },
-  {
-    text: "MT (예시)",
-    onClick: () => console.log("MT 예시"),
-  },
-];
