+++ src/components/meeting/create/entities/ParticipationChips.tsx
@@ -1,5 +1,5 @@
 export type ParticipationChipsProps = {
-  avatar: string; // 프로필 이미지 ur
+  avatar: string;
   onClick?: () => void;
   selected?: boolean;
   children?: React.ReactNode;
@@ -29,30 +29,15 @@ export default function ParticipationChips({
         className="w-6 h-6 rounded-full object-cover"
       />
 
-      {/* 이름 */}
       <span className="text-gray-700">{children}</span>
 
-      {/* 닫기 버튼 */}
       {selected && (
         <button
           type="button"
           onClick={onClick}
           className="ml-1 text-gray-400 hover:text-gray-600 focus:outline-none"
         >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            className="h-4 w-4"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M6 18L18 6M6 6l12 12"
-            />
-          </svg>
+          X
         </button>
       )}
     </div>
