+++ src/mocks/meeting/create/templates.ts
@@ -1,0 +1,18 @@
+export const TEMPLATES = [
+  {
+    text: "안건 & 논의사항.",
+    onClick: () => console.log("안건 선택"),
+  },
+  {
+    text: " 녹음 파일로 회의록 작성하기",
+    onClick: () => console.log("녹음 파일로 회의록 작성"),
+  },
+  {
+    text: "신입생 OT (예시)",
+    onClick: () => console.log("신입생 OT 예시"),
+  },
+  {
+    text: "MT (예시)",
+    onClick: () => console.log("MT 예시"),
+  },
+];
